Added loading bar
With the firebrick red loading bar and spinner-icon the user will be
distracted and informed of  the slight wait.
This should make the UX better. A cache isn't appropriate at this time.
The big wait is usually on initial load of check register with > 5000
entries. Cache is no help there. Refresh only occurs on change to DB;
cache doesn't help there either. Closes #6